WIP: merging crubyjruby specs
